--- src/main/java/roomescape/repository/MemberRepository.java
@@ -1,14 +1,0 @@
-package roomescape.repository;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import roomescape.domain.member.Email;
-import roomescape.domain.member.Member;
-import roomescape.domain.member.Password;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByEmailAndPassword(Email email, Password password);
-}
