+++ src/main/java/roomescape/service/WaitingService.java
@@ -5,11 +5,11 @@ import static roomescape.domain.reservation.ReservationStatus.WAITING;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.member.Member;
+import roomescape.domain.member.MemberRepository;
 import roomescape.domain.reservation.Date;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.ReservationTime;
 import roomescape.domain.theme.Theme;
-import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
